Update R examples
Reorganize R directory to use a single renv dir, which requires being
in a single project.

Make the R directory the project and be ok with the package being a
subdir of the project, despite all of the frantic warnings from
usethis.

Add a script to create a data report from dataReporter (formerly
dataMaid).